Add user hook
