+++ pages/index.tsx
@@ -4,10 +4,8 @@ import FullSpinner from 'lib/components/full-spinner'
 import SelectRegion from 'lib/components/select-region'
 import withAuth from 'lib/with-auth'
 
-function SelectRegionPage() {
+export default withAuth(function SelectRegionPage() {
   const {data} = useSWR('{regions {name, _id}}')
   if (!data) return <FullSpinner />
   return <SelectRegion regions={data.regions} />
-}
-
-export default withAuth(SelectRegionPage)
+})
